Cast of a size_t into an int for string formatting in a log error message. Assertion was raised.

+++ vroomgis/src/vrdriver.cpp
@@ -94,7 +94,7 @@ vrDRIVERS_TYPE vrDrivers::GetType(const wxString & extension) {
 	}
 	
 	if (extension.Len() < 3){
-		wxLogError("Only %d character in the extension (need 3 character)", extension.Len());
+		wxLogError("Only %d character in the extension (need 3 character)", (int)extension.Len());
 		return vrDRIVER_UNKNOWN;
 	}
